requirements.txt
boto3==1.24.2
numpy==1.23.5
openai==0.27.8
pandas~=1.3.5
pytz==2023.3
requests==2.31.0
SQLAlchemy==1.4.49
botocore<1.28.0
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
